Convert Davify script to be a bit more professional
